+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
-## MusicPlayerPlus-3.0.1r2 (2023-01-11)
+## MusicPlayerPlus-3.0.1r2 (2023-01-17)
 * Use external `mppcava` package
 * MusicPlayerPlus package is now architecture independent
+* Improved cover art display
+* Improved platform detection
 
 ## MusicPlayerPlus-3.0.1r1 (2023-01-09)
 * Arch Linux modular installation fixes